Cleaned up and removed old command-line code
