+++ src/command_line/builtin_commands/__init__.py
@@ -1,6 +1,3 @@
 from command_line.builtin_commands.world_commands import WorldCommand
 from command_line.builtin_commands.box_commands import BoxCommand
 from command_line.builtin_commands.operation_command import OperationCommand
-from command_line.builtin_commands.base_commands import (
-    VariableCommand, PopModeCommand, ReloadCommand
-)
